Replace pseudo-Makefile with real Makefile
This required me to get rid of spaces in the raw data from King and Lowe (2008). Rather than manually replace the spaces, I decided to go ahead and write the script to download these data directly from Dataverse and save them with better names.

I switched back to VB to fit the model.